requirements.txt
keyboard==0.13.5
python-vlc==3.0.20123